corrected opengraph. page.tsx now are server components
