+++ app/layout.tsx
@@ -49,4 +49,3 @@ export default function RootLayout({
     </html>
   );
 }
-2;
