+++ app/membership/membership.tsx
@@ -1,0 +1,107 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import Image from "next/image";
+
+export default function Membership() {
+  return (
+    <div className="flex flex-col min-h-screen bg-gray-50 dark:bg-gray-900">
+      <main className="flex-1">
+        <section className="w-full pt-12 md:pt-24 lg:pt-32">
+          <div className="container px-4 md:px-6">
+            <div className="grid gap-6 lg:grid-cols-2 lg:gap-12 xl:grid-cols-2">
+              <Image
+                alt="Lifetime Membership"
+                className="mx-auto aspect-square overflow-hidden rounded-xl object-cover object-center sm:w-full"
+                height={400}
+                src="https://assets.pox.me/memberships/membership.png"
+                width={250}
+              />
+              <div className="flex flex-col justify-center space-y-4">
+                <div className="space-y-2">
+                  <h1 className="text-3xl font-bold tracking-tighter sm:text-5xl xl:text-6xl/none">
+                    Lifetime Membership
+                  </h1>
+                  <p className="max-w-[600px] text-gray-500 md:text-xl dark:text-gray-400">
+                    Get unlimited access to all our premium features with a
+                    lifetime membership.
+                  </p>
+                </div>
+                <ul className="grid gap-2 py-4">
+                  <li>
+                    <CheckIcon className="mr-2 inline-block h-4 w-4" />
+                    50% Discount in the premium features.
+                  </li>
+                  <li>
+                    <CheckIcon className="mr-2 inline-block h-4 w-4" />
+                    Priority customer support
+                  </li>
+                  <li>
+                    <CheckIcon className="mr-2 inline-block h-4 w-4" />
+                    Free access to all future updates
+                  </li>
+                </ul>
+                <div className="grid gap-6 lg:grid-cols-2 lg:gap-12 xl:grid-cols-2">
+                  <h1 className="text-3xl font-bold tracking-tighter sm:text-5xl xl:text-6xl/none">
+                    $300
+                  </h1>
+                </div>
+              </div>
+            </div>
+          </div>
+        </section>
+        <section className="w-full py-12 md:py-24 lg:py-32 bg-gray-100 dark:bg-gray-800">
+          <div className="container px-4 md:px-6">
+            <div className="flex flex-col items-center justify-center space-y-4 text-center">
+              <div className="space-y-2">
+                <div className="inline-block rounded-lg bg-gray-100 px-3 py-1 text-sm dark:bg-gray-800">
+                  Payment Options
+                </div>
+                <h2 className="text-3xl font-bold tracking-tighter sm:text-5xl">
+                  Choose your preferred payment method
+                </h2>
+                <p className="max-w-[900px] text-gray-500 md:text-xl/relaxed lg:text-base/relaxed xl:text-xl/relaxed dark:text-gray-400">
+                  We offer multiple payment options including crypto, PayPal,
+                  credit card, and a rental option.
+                </p>
+              </div>
+              <div className="flex flex-wrap justify-center gap-4">
+                <Button disabled className="w-full sm:w-auto">
+                  Pay with Crypto
+                </Button>
+                <Button disabled className="w-full sm:w-auto">
+                  Pay with PayPal
+                </Button>
+                <Button disabled className="w-full sm:w-auto">
+                  Pay with Credit Card
+                </Button>
+                <Button disabled className="w-full sm:w-auto">
+                  Rental Option
+                </Button>
+              </div>
+            </div>
+          </div>
+        </section>
+      </main>
+    </div>
+  );
+}
+
+function CheckIcon(props: any) {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <polyline points="20 6 9 17 4 12" />
+    </svg>
+  );
+}
